xpdf: fix compilation
It is too bad that it is prohibitively expensive to set up
special-purpose MSYS2 versions to build *just* one package. That would
preven stupid mistakes like this one, where a dependency (libpaper) was
installed in this developer's setup, but is not declared as a dependency
(and indeed, should not be a dependency).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>