curl: fix memory corruption in debug mode, and a WinSSL issue
We never paid attention to curl's new features, apparently, and
therefore we completely missed when curl introduced the ability to
override the allocator, or debug memory allocations.

As a consequence, our code breaks horribly when we compile it in debug
mode, when the code tries to release memory in url.c that was allocated
in pathtools.c (because pathtools.c was happily oblivious to the way
curl wants the malloc() to be performed).

While at it, fix the dozens of compiler warnings, quite possibly fixing
some memory issues on 64-bit where possibly only the lower 32 bits of a
pointed would have been used.

Also, backport the fix for the new feature that allows Secure Channel
(called "WinSSL" by curl) to use a user-provided ca-bundle.crt file:
without the fix, the ca-bundle.crt cannot have Windows line endings.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>