Add two helper macros to the PKGBUILD templates.  apply_patch_with_msg applies a patch ensuring a backup is made and with an output message showing the patch being applied to help with debugging.  del_file_exists delets files and is meant for where a patch creates new files.  Those added files may not be deleted if the PKGBUILD is reran causing patch to fail saying it's already been applied.
Fix the setup-py template to use a _dtoken variable for part of the URL in pypi.  Those use a download token for each software package and is unique to EVERY version.  Also, use python${pver}-build-${CARCH} consistantly.  The CARCH allows us to have for binary versions so that they could be examined if a build problem were to occur and speed up the build.