Do not force-set HOME variable, let nsswitch decide
Previous versions of Git/msys1 did not check for the
correct location of the user profile on Windows, thus
requiring the HOME variable to be set. Since Git 2 nsswitch
is used to identify the correct location of the user profile.

For non-roaming domain userprofiles, where USERPROFILE still
remains on drive C:\, but with an additional home drive set,
this line breaks search for the SSH key, since nsswitch returns
the correct results, but HOME is used instead, since it has been
already set by start-ssh-agent.cmd.

Signed-off-by: Nathan-J. Hirschauer <nathan-jedidja.hirschauer@draeger.com>