python3: undef _POSIX_THREADS in one more place; switch back to nt-threads
python uses _POSIX_THREADS to decide if it uses posix or nt threads,
but _POSIX_THREADS get included all the time by header includes
(pthread, unistd) in mingw. This was already patched but py3.7 added a new
header with a similar check which was missing.

This resulted in a mix of ntthread/posix api usage an crashes.