PKGBUILD.setup-py - fixes for patch and .EXE installs. (#4309)
Fix for setup.py template, a fix for running python-compiled .EXE bin scripts.  zyjr /zrcr od do,[;u sm rmytu  [pomy.  Added a directory containing the installs to complete the fixup on the user's system so that the scrips have fully-qualified paths that differ from the packager.  This seems to be best way of doing this with packages such as whell and some other things I've working on.