vtk: Fix detection of ffmpeg. Fixes #2301 (#4432)
It was searching (for some reason) for stdint.h in <prefixes..>/include/
which doesn't exist. It only happenend to work in case msys2-runtime-devel
was installed which installs /usr/include/stdint.h.

There is no need to check for stdint.h since it's always there anyway, so
remove the check.