python-setuptools: unvendor dependencies. Fixes #4623
Since pip got unvendored in #4379 the Python classes of system packages
and vendored packages clash. This removes the vendored packages from
setuptools so only system dependencies are used in both pip and setuptools.

The downside of this is that it's not officially supported by upstream
and this creates cyclic dependencies and makes bootstrapping harder.

If this turns out to be a problem we should probably revert this
and #4379.