Add mingw-w64-gumbo-parser package (#4652)
* Add mingw-w64-gumbo-parser package

The gumbo parser library is one of the dependencies for the talkatu
library, which is a dependency of the current development version
of Pidgin 3. This patch adds the PKGBUILD for the gumbo parser
library. It's ported from Arch's package which is maintained by
Balló György <ballogyor+arch at gmail dot com>.

* gumbo-parser: Append "(mingw-w64)" to package description

* gumbo-parser: Use $pkgver in favor of hardcoding source url version