git: allow for -municode
When compiling the Git wrapper, we reuse Git's `CFLAGS` and `LDFLAGS`.
The latter recently acquired the `-municode` flag, and we were
unprepared to compile with that, as we only defined a `main()` function.

Let's define a `wmain()` function instead when we are compiled with
`-municode` (which implicitly sets the `UNICODE` constant).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>