git: avoid strange 32-bit problems with git-bash.exe
In v2.19.2, it was reported that git-bash.exe does not start correctly.
Turns out that the option -Wl,--pic-executable,-e,mainCRTStartup somehow
produces incorrect code now, and it is not super critical to keep that
option, so let's just ignore the issue and drop that option when linking
our Git wrapper executables.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>