git: avoid rewriting `LDFLAGS` in `mingw-w64-git.mak for 32-bit builds
In deb0395d0 (git: avoid strange 32-bit problems with git-bash.exe,
2018-11-22), we tried to address the problem where 32-bit `git-bash.exe`
would simply segfault instead of starting.

However, it turns out that incorrect testing misidentified this as the
fix, and we later committed the correct fix: 4b5e86d8f (git: really
avoid strange 32-bit problems with git-bash.exe, 2018-11-22).

Now, the former fix not only turns out to be unnecessary, but also to
cause `git.exe` to be re-linked in the `strip` target, as we are using
`mingw-w64-git.mak` for that target, and in that file, `LDFLAGS` was
rewritten (incorrectly, as it turned out, but still resulting in a
change of leading whitespace in `ALL_LDFLAGS`).

Let's just revert that unneeded (and unwanted) change.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>