curl: downgrade from HTTP/2 to HTTP/1.1 when needed (e.g. for NTLM)
NTLM is incompatible with HTTP/2, and most (or all) of Kerberos, too. In
such cases, it is a problem that cURL v7.62.0 changed the default to use
HTTP/2 for https:// connections whenever servers indicate support for
it.

This is a backport of https://github.com/curl/curl/pull/3345 and
https://github.com/curl/curl/pull/3349, so that Git for Windows v2.20.0
will have those fixes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>