git: only hide console if explicitly asked for
In 514aa5b00 (git: fix --hide in git-wrapper.c, 2018-12-06), we tried to
ensure that `git-bash.exe --needs-console --hide` works as expected: it
allocates a console but hides the window.

However, the way we implemented this broke `git-bash.exe` without
options, as it hit the very same code path as the `--hide` route (nobody
asks `git-bash.exe` to show the console that it allocates).

Let's fix this by turning the `show_console`/`no_hide` variables into
tristates: do-nothing/false/true (via -1/0/1). And then we really do not
try to show or hide the window if neither `--hide` nor `--no-hide` are
in effect: we simply let `CreateProcess()` do what it thinks it should
by default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>