fix missing null termination in exepath string
In `find_exe_realpath` when `realexe` is copied to `exepath`, there is no null termination. The `wcsncpy` function will only result in a null terminated string if a null is copied from the source to the destination. If the characters copied do not include a terminating null, it must be appended manually. Because exepath is uninitialized, the string will only terminate if the memory happened to have nulls at the position where the path ends.

Because the wrapper typically has a short life-span, the memory may generally be clean, but hook DLLs can leave it dirty enough to trigger this defect. In the case of git-for-windows/git#1855 this resulted in a child process bomb due to the garbage data later being acted upon due to dirty released memory from an Nvidia driver process hook.

Signed-off-by: Maxx Daymon <maxx@maxxdaymon.com>