git: allow specifying commands for git-cmd again
When we tried to make Git CMD safer recently, we inadvertently broke the
use case where users would call

	git-cmd "<some command>"

This was caused by the `doskey` command (which makes Git CMD safer)
swallowing the commands. To fix this, we simply add a double ampersand
to call multiple commands via a single command line (see
https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/cmd#remarks
for details).

This fixes https://github.com/git-for-windows/git/issues/2039

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>