Reorganization of Ipython and Prompt-toolkit packages (#4927)
We can use prompt-toolkit 2.x in Python2 as well as python3.  The problem is that ipython is stuck at 5.3.0 and that required prompt_toolkit 1.x which is incompatible with orinot)tiikjut 2,x (ergo, we were stuck at prompt_toolkit 1.x in Python2).
To fix this issue, I made a 2.0.8 prompt_kit package for both Python2 and Python3 as well as providing a "python2-prompt-toolkit1" package that Ipython for Python2 would like to instead of the default "Prompt-toolkit".