python3: fix syntaxerror in sysconfig for some builds
The build scripts does basic sed replacement to make things relocatable
and fails when strings are wrapped (depending on the path lengths).

Disable wrapping when writing out the file to prevent that.

Ideally we would not use sed and write out the correct values in the first place,
but that's for another time.

Fixes #5048