Use current path prefix to find Python script files to modify
When using pacman to install packages, the --root option can be set to
install them to an alternative location. The current install scripts
don't take this in to account which causes no such file or directory
errors. This modifies the sed command to use the full prefix path
when renaming the shebang.

Signed-off-by: Dan Yeaw <dan@yeaw.me>