python: better compatibility with unpatched setuptools. Fixes #5155
We currently patch setuptools itself to not import distutils.msvc9compiler,
which fails because it fails to detect the msvc version and falls back to an unsupported
version 6 and raises.

This doesn't help much in case setuptools is installed through pip which doesn't contain
the fix.

This fixes msvc9compiler.py instead to not fail on import but at the compiler instance creation,
i.e. the point where the version is actually used. The setuptools fix remains for now to make
updates easier.

Fixes #5155