cmake tweak FindSWIG command such that users can set a search path for swig
As mingw is not typically installed in a path that's hard-coded into cmake, the original find
command would not find swig. By allowing the user to set search paths (that is, by removing
the NO_CMAKE_FIND_ROOT_PATH flag), a user can tell cmake where to look for swig.