gettext: fix tools relocation support
relocate() in gnulib-lib needs INSTALLDIR and INSTALLPREFIX to be
Windows paths so it can relocate the input.

Fix by removing INSTALLPREFIX from MSYS2_ARG_CONV_EXCL so it gets
converted when passed to the compiler and pass it to gnulib-lib,
and convert INSTALLPREFIX in gettext-tools/config.h manually.

Fixes #4392 (for real, hopefully)