Split client tools and libs out of postgresql package
This greatly reduces the amount of files and dependencies to be installed
when no server libraries are needed.

The new package is called "postgresql-client".
Package "postgresql" installs "postgresql-client" implicit, so that the
package is fully backward compatible.

For some reason gcc links to libxml2.dll, when linker is called with -lxml2 ,
although --as-needed is specified. Unfortunatelly there seems to be no easy
way to disable -lxml2 for client libraries only, so that the package
"postgresql-client" depends on libxml2, although it has no use there other
than fulfilling this dll dependency.