Add back python2-setuptools
This adds back a python 2 version of setuptools but with all dependencies vendored
(the upstream default..). To allow users to install Python 2 packages we no longer
have in the repo etc which should help them transitioning.

This also adds conflicts between py2/3 setuptools because of the esy_install rename.

See #4993