gtk3: only install low level keyboard hook if GTK_CSD=1
The hook is used to intercept keyboard combos in order to provide
AeroSnap-like functionality for CSD windows.

However it causes significant performance issues, especially on
startup of gtk3 apps (those are system-wide and *not* limited to the
gtk3 app itself). Performance can be continuously degraded to any
console application that is attached to the gtk3 app, e.g. debuggers.

Related issues:
- https://gitlab.gnome.org/GNOME/gtk/issues/2015
- https://gitlab.gnome.org/GNOME/gtk/issues/1082
- https://gitlab.gnome.org/GNOME/gtk/issues/1033

As the hook is completely useless for non-CSD windows, only install
it if CSD is explicitly requested (environment variable GTK_CSD=1).