sqlite3: add sqlite extensions to package
sqlite extensions are shared libraries loadable at run-time enhancing
sqlite's SQl code.

The source code of each of the sqlite extensions is located in directory
'ext/misc' of the source code package.  A hand-written Makefile, adjusted
using a project's autoconf tool (config.status), is generated in the
extensions build folder.

The extensions (together with each of the source c-files containing the
extension documentation in the file header) are installed in the directory
'${prefix}/share/sqlite/extensions'.

* PKGBUILD:
  - increase pkgrel
  - add build and install routines for sqlite extensions.
* Makefile.ext.in: add file.
* README.md.in: add file.