gcc: hack to fix precompiled headers.
At least cc1 and cc1plus need to NOT be ASLR, or PCH won't work.  I am
not sure about any other backend, so err on the side of caution and
remove the dynamicbase flag from all of them.

NOTE: peflags is provided by the rebase package (a cygwin tool, and part
of base), and does not seem to support the high-entropy-va flag.
Fortunately it seems Windows ignores the high-entropy-va flag if
dynamicbase is not set.