Switch gmp to install shared headers.
This requires mpfr, mpc, and isl to be built shared rather than static,
as their configure scripts check that they are being built in the same
static/shared configuration as the gmp headers.

This allows gcc to no longer require the --default-image-base-low flag
on x86_64.  Unfortunately, until gcc is rebuilt with proper linkage, the
gmp dll requires that flag, or the old gcc crashes.

Fixes #6983.