curl: preserve UNC paths when normalizing/simplifying
The code to make cURL relocatable simplifies/normalizes paths, e.g.
reducing multiple forward slashes to a single one.

This is okay in general, but at the beginning of the path, a double
slash has a special meaning: it denotes a UNC path of the form
//<host>/<path>. It would therefore be a mistake to reduce those first
two slashes to a single one: otherwise it would refer to the absolute
path relative to the current directory's drive.

Let's preserve the double slashes.

Reported-in: https://github.com/git-for-windows/git/issues/3266
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>