Allow building conflicting packages by the same CI
This PR change the way ci build packages.
Previously the ci sort the packages by dependency and then build them
one after another. The upside of that approach is to be able to build dependent packages
within the same job, the downsides is that approach makes it impossible
to build conflicting packages within the same job, and also doesn't
offer a clean environment to build the independent packages.

With this approach we uninstall every package after a successful build
while adding a local (ci) repository from where the following packages
can install their dependencies.
This approach will allow to build conflicting packages within the same
job and offering a clean environment to build the independent packages,
so we can discover missing make-dependencies.