CI: track package directory in ci-get-build-order.py
If a package being built depends on a split package, ci-get-build-order would return the split package name, not the original directory name it started with, which may be different.  I don't know if this happens in MINGW-packages, but it was the case with curl/libcurl in MSYS2-packages.

Fixes MSYS2-packages#2634