xpm-nox: fix #8548 XPM writes invalid X Pixmap with path as C identifier
The Windows path characters : and \ were not filtered out when writing
the C identifier that is part of the XPM image.
Besides that writing the complete path inside an image can
lead to a C identifier which is too long.

So let's just remove the path and only leave the filename.