ncurses: define NCURSES_STATIC by default again
this was removed in https://github.com/mirror/ncurses/commit/04d942c3d98cf0a929c6afb17be8c10d4ae39af0

We only provide static builds right now and the .pc file doesn't
include a Cflags.static so no users know that this needs to be set
or that only static builds are available.

So revert this for now.

In the future we could look into adding shared builds and define a Cflags.private
for static builds. Then this patch could be removed again.

Fixes #10312