CI: only install the C compiler instead of the whole toolchain group
See https://github.com/msys2/MINGW-packages/discussions/10506

This is an intermediate step. It results in CI installing fewer packages
by default before building, but makepkg-mingw will not complain since
it only checks for the C compiler when checking for the group.

If this works out we can update the docs and remove the install alltogether.