mingw-w64: update to 10.0.0
I would have updated the pkgver to 10.0.0 even, since this is the commit
officially tagged as v10.0.0, but this package is updated against git
commits so often that tweaking the pkgver function for one version
didn't seem worth the effort.

Take this opportunity to switch pkgver to using git describe, as doing
so when the more significant version numbers don't change would make the
new version compare lower.

The --long argument and sed pattern for git describe comes from the
CMake-git PKGBUILD template.