openjepg2: Reapply stdcall for all library type
This patch was disable in 8edfe204d937edf31b0310c6195de5a337c270b9 commit.
But it fixes static linking with 32 bit ffmpeg. Previously, the stdcall
decoration was only enabled in shared libs but not in static libs. So,
when OPJ_STATIC is enabled the ffmpeg configure fails to check function
without stdcall decoration in 32 bit environment.