binutils: Move include and libs back to native directories
They were moved to a subdirectory in 047d1d662f (binutils: resolve
conflicts with headers and libs with gdb, Dec 2013). Since then, gdb and
binutils repositories were merged, and gdb does not install these headers
and libs at all, so there is no conflict.

Linux systems have bfd.h, libbfd.a and libiberty.a in include and lib, and
not in a subdirectory.

Also, placing them in a subdirectory requires the users to specify the lib
directory, and have non-portable #include statements (binutils/bfd.h on
MinGW, else bfd.h).

Keep the original locations as symbolic links to maintain backwards
compatibility.