gtk2: remove gtk-update-icon-cache
We want to provide this command via the gtk-update-icon-cache package
which is currently built via gtk3 (maybe gtk4 in the future), so remove it
in gtk2 to avoid conflicts.

Also remove the dependency on the gtk-update-icon-cache package, since
gtk2 doesn't install any icons, so no need to trigger the cache hook.

See #12527