crt: replaces/conflicts libssp
Up until now a full upgrade would result in libssp being updated to
a version that only has the DLLs left, but the packages got removed from
the repo now, so users upgrading after the removal will get a conflict
with the now no-longer-in-the-repo package.

Replace/conflict with it so it gets automatically uninstalled on a system upgrade.

Fixes #13534