waf: default compilers to the corresponding package MinGW system
Currently, when running mingw64/waf on a default install, waf will use
the default C and C++ compilers for the win32 platform configured in
waf, which are [msvc, gcc, clang] and [msvc, g++, clang++].

This means that waf tries to use msvc first when compiling waf projects.
This is not what the user want: instead, it should use the appropriate
MinGW system compiler, eg gcc on mingw64.

This currently requires the user to work around the issue by
systematically adding --check-c-compiler to point waf to the appropriate
compiler depending on the MinGW system, eg gcc on mingw64.

This patch fixes the issue by changing the default C and C++ compiler
order in waf, to move the corresponding MinGW compilers to the head of
the list.

(We could have set "gcc, msvc, clang" for mingw* platforms and "clang,
msvc, gcc" for clang* platforms, but setting "gcc, clang, msvc" works
well enough in all cases.)

This is done by using the preamble build-time option of waf which
enables users to add a small Python preamble that is evaluated for every
project before their wscript file. In this preamble, we update the
c_compiler and cxx_compiler values.