openssl: fix modulesdir relocation
openssl 3 added a new path from which DLLs can be loaded at runtime,
currently the "legacy.dll".

This code path was missing the relocation support and looking for
the CI build path instead, resulting in code that tried to load
the legacy providers to fail (examples: python-cryptography, transmission)

It can also be reproduced via:
   openssl list -provider legacy -providers

Add relocation support to MODULESDIR to fix this.