rust: set _bootstrapping=no for all architectures (not just aarch64)
when _bootstrapping=no, the rust build is configured to use the local rust for stage 0.
stage 0 will use the installed MSYS2 rust (must be the latest version...).

when _bootstrapping=yes, the build is configured to download and extract the previous rust version.
in this case, the official rust build will be used which kind of defeats the point (?)