dxil-redist: new package
Contains copies of the dxil.dll redistributable libraries from
the Windows SDK.

As dxil.dll is not normally present in system installations, we
have to ship it ourselves.

Note that another official distribution channel for dxil.dll is
https://github.com/microsoft/DirectXShaderCompiler/releases, but
we avoid using that because the distributed modules are dependant
upon the Visual C++ runtime. Instead, the modules from the Windows
SDK are built with the Visual C++ runtime linked statically.