Add findutils
Patch 001 adds the poll and windows-stat-inodes gnulib modules. The
poll module provides the poll functionality that is absent on
mingw. windows-stat-inodes ensures the -inum option of find is working
properly. The modules have been taken from the same gnulib commit that
is linked in the findutils 4.9.0 release (v0.1-5153-g6ef3d78333).

Patch 002 modifies the configure script to allow using the mountlist
module on mingw. The code doesn't compile without it but it also
doesn't seem broken. The mountlist module also specifies that using
the header is fine on unsupported systems.

Patches 003, 004, 006, 007, 010 and 015 have been adapted from Eli
Zaretskii's ezwinports:

https://sourceforge.net/projects/ezwinports/files

Patches 005, 009, 013 and 014 disable features that had not been
implemented yet in the 4.2.30 findutils release that ezwinports is
based on. I also don't know how I'd implement them; they are all based
on functionality that is not available on mingw as far as I can tell.

Patches 008 and 011 are fairly straightforward compilation error
fixes.

Patch 009 tries to address the differences in how the PATH environment
variable is separated on Windows.

Patch 016 fixes linking on clang builds.

Note that not all of the tests pass. I could provide patches for some
of them if this is preferable, but for others I think there's no use
case on Windows. I keep a write-up here:

https://github.com/cyrilarnould/findutils/blob/2ccb0be2c1c70d791f051d24b9dba73e64bdb446/README-mingw