libilbm: Add package for GIMP
GIMP has a new plug-in for importing Amiga IFF images which depends on libilbm. This library depends on libiff which was just added to the MSYS2 package repository.
This patch attempts to add the libilbm package to MSYS2 to finish adding support for the GIMP IFF plug-in in Windows builds.