libva-utils: add package
As of libva version 2.18.0 and mesa 23.1.1 it's required to set
the following environment variables to be able to work with vaon12
driver:
  $ export | grep LIBVA
  declare -x LIBVA_DRIVERS_PATH="/mingw64/bin/"
  declare -x LIBVA_DRIVER_NAME="libvaon12"

This will be fixed in the next versions of libva and mesa - patches
for that already landed in upstream versions of these components.

With above environment variables set user should see something like
this (output depends on actual HW capabilities, below is a snapshot
from Interl Alderlake running Windows 11):

  $ vainfo
  Trying display: win32
  libva info: VA-API version 1.18.0
  libva info: User environment variable requested driver 'libvaon12'
  libva info: Trying to open C:/msys64/mingw64/bin/\libvaon12_drv_video.dll
  libva info: Found init function __vaDriverInit_1_18
  libva info: va_openDriver() returns 0
  C:\msys64\mingw64\bin\vainfo.exe: VA-API version: 1.18 (libva 2.18.2)
  C:\msys64\mingw64\bin\vainfo.exe: Driver version: Mesa Gallium driver 23.1.1 for D3D12 (Intel(R) Iris(R) Xe Graphics)
  C:\msys64\mingw64\bin\vainfo.exe: Supported profile and entrypoints
        VAProfileVP9Profile0            : VAEntrypointVLD
        VAProfileVP9Profile2            : VAEntrypointVLD
        VAProfileAV1Profile0            : VAEntrypointVLD
        VAProfileNone                   : VAEntrypointVideoProc

To enable more features (h264 and h265 decoder/encoder and others)
it's required to build mesa with corresponding capabilities. That's
a subject for other patch in msys2 mesa build.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>