Fix clang build by disabling an error treated as a warning by gcc
Clang generates a compiler error against code which gcc treats as a warning. This change disables that error (only for clang) so the build can succeed. The code needs to be fixed (has been raised with the project), but this allows the build to succeed so downstream consumers (e.g. Gramps) can use the package.

Bumped up the package release.