Re-add mingw libtool variant
It was removed in #10338 because it wasn't really used and the
msys one seemed good enough for autotools (via libtoolize), and
all packages using it were really just depending on libltdl.

Turns out some mingw packages use libtool directly and the msys
one has various things hardcoded for that environment, resulting
for example in wrongly named .dlls.

This adds back the mingw libtool with all patches synched from the
msys one and the Windows paths fixed up in the libtool script.
Also add it as a dependency to autotools, so we avoid sometimes
the msys one being used and sometimes the mingw one, which would
make it harded to debug.

Ideally this doesn't change anything for existing packages and
just fixes the ones using libtool directly, for example
mingw-w64-libvterm and mingw-w64-libtommath.

There is also a libtool integration test being added now in
https://github.com/msys2/msys2-tests/pull/45