scrcpy: Help scrcpy find the server and icon file
Scrcpy uses `PREFIX+"/share/scrcpy/scrcpy-server"` as the default server
path, and `PREFIX` is `MINGW_PREFIX` (determined at build time). Before
using the server file, scrcpy checks to see if the path to the server
file is legal. Since the default server path is an MSYS2 path and not
a Windows path, the path must not be legal, and scrcpy will not be able
to find the server file.

Scrcpy implements a function (`sc_file_get_local_path`) that gets the
program directory and allows you to concatenate a string after the path.
I used it to splice the program directory with the relative path to the
server file as the server path, which solved the problem.

Same for icon file.