binutils: Update to 2.42
Remove all backports which are included in this new release.

Remove specify-timestamp.patch, looks like this was implemented upstream in:
https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=6badd1020f5bebd3f
and it seems to work here.

Work around binutils linking against the system libiberty instead of
the included one.

Also binutils now links against libintl, so add gettext as a dependency