Add xmlto
This is like the msys xmlto but calls mingw tools where possible.
Using git as a source since there are various build fixes there
that haven't been released.

The man page output is garbled, but I'm not sure if this is a
xmlto or a problem with the other tools.

Fixes #20018