rust: update to 1.77.1 (#20397)
* rust: update to 1.77.1

* libc++: enable frame apis in libunwind.

The bootstrapping hack used by the rust package on clang subsystems
(namely, substituting compiler-rt and libunwind for libgcc and using the
official gcc-based binaries to bootstrap) attempts to link to
__register_frame_info and __deregister_frame_info on i686.  Enabling
this option brings them back.

* rust: backport rust-lang/compiler-builtins#575

Also, fix 0011 patch to Cargo.lock (it was adding the embed-manifest
dependency to the wrong crate due to fuzziness), and
vendor embed-manifest.  This allows CLANG32 to build with
--enable-vendor, which is required to use our now-patched version of
compiler-builtins.  Note the 67c1c0a71a204c089ddae4aec21ec75aa778c11b
commit was *not* the version merged upstream, but *is* the version that does
not fail on either MINGW32 or CLANG32.

* rust: enable uac patch for all clang prefixes.

It turns out to be required on i686 *and* aarch64, but shouldn't hurt on
x86_64 either.

---------

Co-authored-by: Jeremy Drake <github@jdrake.com>