CI: fix minor issue.
The test to see if the packaegs array is empty was not correct.  When
there was more than one entry, it would result in the error message:
 ./.ci/ci-build.sh: line 113: test: too many arguments
Instead, check if the number of items in the array is 0, which is
similar to what is done with the binaries array below (though it checks
if the number of items is not 0).