Add cmake-bootstrap
A cmake variant that doesn't depend on any external dependencies and
can be used for bootstrapping an env, and building a full cmake.

The problem with cmake is that it depends on various libraries such
as curl, libarchive, libuv etc for which we want to use cmake instead
of autotools (native, faster, better supported on Windows).

These cycles prevent automated builds in case one of the dependencies
has an ABI break, or when bootstrapping a new environment.

Other distros avoid this by partly vendoring some dependencies, using
autotools for certain dependencies to avoid cycles, and not including the
cmake GUI tools. We'd prefer to use cmake over autotools everywhere if
supported though, and provide a full cmake installation to users.