qt5: restore some MINGW32 packages
Following a discussion on Discord, see:
https://www.discord.com/channels/792780131906617355/794220101741838401/1283064557660803083
brought about by https://github.com/msys2/MINGW-packages/commit/a4f0579e14509d27927a94498c35e1ccddf29bf5
I suggested to restore for a while a couple of Qt5 packages that the
[Mudlet](https://github.com/Mudlet/Mudlet) project I code for uses for its
32-Bit Windows builds until we have migrated as many of our Windows users
up to 64-Bit builds as possible.

I was hoping that one of the MSYS2+Mingw-w64 team would have been able to
find the time to make this change but after a couple of weeks of nothing
happening I have decided to try and do it myself.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>