openssl: Update to 3.4.0
Refresh 002-relocation.patch.

The whole openssl dir code was changed upstream and now has everything
in one place with the option to read the dirs from the Windows registry
if built with OSSL_WINCTX defined.

The registry parts are from what I see namespaced by version and written
by the official installer, so we probably don't want to use that to avoid
any conflicts with the official installation.

The upside though is that we need to add relocation in fewer places now,
and even the "openssl info" commands will use the new functions to now
show proper Windows paths instead of the configure time cygwin ones.

Besides that there shouldn't be any functional change.

and -j1 for make install to avoid it failing