mkvtoolnix: avoid configuration error
The boost library `libboost_system-mt.dll` was an empty dummy that is no
longer being distributed since boost is built with cmake (see #22040).

Configuration fails if no `boost-system` library is found. Work around
that configuration error by pointing to an existing library in the
respective configure flag. Use a library that is going to be used anyway.